Add the callback attribute
A new attribute `callback` is added, to be used as `#[ramhorns(callback = "some_fn")`.
For a field of type `T`, it allows one to pass a function of type
`(&T, &mut E) -> Result<(), E::Error)>` where `E: Encoder` to customise
the rendering of the field.

`#[md]` attribute is replaced with `#[ramhorns(md)]` for clarity and consistency
with other attributes.
`render_cmark` method is deleted as it no longer serves a purpose (a callback
is used instead).