Add the ability to extend the template collection from another folder
Adds functions `extend_from_folder` and `extend_from_folder_with_extension`
to include templates from another folder in the collection.
If there is a file with the same name as a previously loaded template or partial,
it will not be loaded.
Also, this should fix the issue maciejhirsz#53 (the name of the partial which was not found
should always be returned).