Allow nesting sections in tags
It often happens that only one field from a subsecion is to be rendered.
For this, one currently has to write `{{#section}}{{field}}{{/section}}`,
which is quite verbose. This PR adds the posibility to write this as
`{{section field}}`, or to nest arbitrary many such subsecions, like
`{{section subsecion field}}`.

Furthermore, it allows opening arbitrarily many sections as
`{{#section subsecion subsubsection}}`, or closing them as
`{{/subsubsection subsecion section}}` (in the order they are closed).

Apart from readability, this should also improve rendering performance a bit,
since unnecessary closing blocks with no HTML data are not added
(`{{section field}}` doesn't require a closing block, the span of the section
is automatically calculated).

On my machine, surprisingly, the parsing benchmark also turns out a bit beter
than before, probably because trimming of whitespace and hashing of fields where
it is not necessary is eliminated.